Some changes to the blinking LED example
Instead of using interrupt on compare match the mode to toggle the OCA1 pin (PB3) was used
by setting the COM1A0 bit of the TCCR1A register. LED was then connected to the PB3-pin.
Results are the same, but the program turned out to be 128-92 = 36 bytes smaller

Also added the data sheet of the Attiny2313